+++ db/pg/backup_restore.go
@@ -1,6 +1,7 @@
 package pg
 
 import (
+	constants "dbdaddy/const"
 	"dbdaddy/db"
 	"fmt"
 	"os"
@@ -8,6 +9,26 @@ import (
 
 	"github.com/spf13/viper"
 )
+
+func DumpDb(outputFilePath string, v *viper.Viper) error {
+	osCmd := exec.Command("pg_dump",
+		"--clean",
+		"--file="+outputFilePath,
+		"--format=directory",
+		"--username="+v.GetString(constants.DbConfigUserKey),
+		"--host="+v.GetString(constants.DbConfigHostKey),
+		"--port="+v.GetString(constants.DbConfigPortKey),
+		"--no-password",
+		v.GetString(constants.DbConfigCurrentBranchKey))
+
+	osCmd.Env = append(os.Environ(), "PGPASSWORD="+v.GetString(constants.DbConfigPassKey))
+
+	osCmd.Stdout = os.Stdout
+	osCmd.Stderr = os.Stderr
+
+	err := osCmd.Run()
+	return err
+}
 
 // creates db & restores its content from given dump
 func RestoreDb(dbname string, v *viper.Viper, dumpFilePath string, override bool) error {
@@ -30,7 +51,6 @@ func RestoreDb(dbname string, v *viper.Viper, dumpFilePath string, override bool
 	// run restore command
 	osCmd := exec.Command(
 		"pg_restore",
-		"--verbose",
 		"--clean",
 		"--if-exists",
 		fmt.Sprintf("--dbname=%s", db.GetPgConnUriFromViper(v, dbname)),
@@ -38,7 +58,6 @@ func RestoreDb(dbname string, v *viper.Viper, dumpFilePath string, override bool
 	)
 
 	osCmd.Stderr = os.Stderr
-	osCmd.Stdout = os.Stdout
 
 	err := osCmd.Run()
 	return err
